requirements.txt
dvc
sklearn
pandas
numpy
numpyencoder
matplotlib
python-dotenv
